Renamed ecal_config to ecal_config_reader
