+++ ecal/core/CMakeLists.txt
@@ -117,7 +117,7 @@ set(ecal_cmn_cpp_src
     src/convert_utf.cpp
     src/ecal.cpp
     src/ecal_clang.cpp
-    src/ecal_config.cpp
+    src/ecal_config_reader.cpp
     src/ecal_descgate.cpp
     src/ecal_event.cpp
     src/ecal_global_accessors.cpp
@@ -220,8 +220,8 @@ set(ecal_service_header_src
 
 set(ecal_cmn_header_src
     src/convert_utf.h
-    src/ecal_config.h
-    src/ecal_config_hlp.h
+    src/ecal_config_reader.h
+    src/ecal_config_reader_hlp.h
     src/ecal_def.h
     src/ecal_def_ini.h
     src/ecal_descgate.h
