+++ ecal/core/src/ecal_globals.cpp
@@ -23,7 +23,7 @@
 
 #include "ecal_globals.h"
 #include "io/udp_init.h"
-#include "ecal_config.h"
+#include "ecal_config_reader.h"
 
 #include <stdexcept>
