+++ ecal/core/src/service/ecal_service_server_impl.cpp
@@ -22,7 +22,7 @@
 **/
 
 #include "ecal_def.h"
-#include "ecal_config_hlp.h"
+#include "ecal_config_reader_hlp.h"
 #include "ecal_register.h"
 #include "ecal_servicegate.h"
 #include "ecal_global_accessors.h"
